feat: Main 컴포넌트 정리
